requirements.txt
zephyr-rtsp
numpy
mss
cv
